fix(sanity): fix issue where context menu wasn't closing when copying to existing release
